Login Unit Finished (v1.1) Refactor login unit code. Added backstage_index as a index interface. Established the connection between these three modules.
