+++ Integration/add_question.py
@@ -634,4 +634,3 @@ class GUI:
 if __name__ == "__main__":
     gui = GUI()
     gui.win.mainloop()
-
